Refactor :gem: :nail_care:
