+++ src/Main/Core/types.ts
@@ -1,0 +1,2 @@
+import { BN, Long } from "@zilliqa-js/zilliqa";
+export type DeployPayload = [{ version: number, gasPrice: BN, gasLimit: Long.Long }, number, number, boolean]
