+++ src/Main/QVoteZilliqa/types.ts
@@ -1,0 +1,4 @@
+import { QVoteContracts } from "../../Utill";
+import { BN } from "@zilliqa-js/zilliqa";
+export type QVPayload = [string, QVoteContracts.Value[]];
+export type CallPayload = [string, QVoteContracts.Value[], { version: number, gasPrice: BN, amount: BN, gasLimit: Long.Long }, number, number, boolean];
