+++ test/exampleWithDecentralizedRegister.ts
@@ -24,37 +24,37 @@ export async function exampleWithDecentralizedRegister(
     */
     const txblock = await zil.blockchain.getLatestTxBlock();
     const curBlockNumber = parseInt(txblock.result!.header!.BlockNum);
-    const gasPrice = await qv.getMinGasHandle(zil.blockchain.getMinimumGasPrice());
+    const gasPrice = await qv.handleMinGas(zil.blockchain.getMinimumGasPrice());
 
     /* Deploy a contract */
     zil.wallet.setDefault(deployerAddress);
-    const contract = zil.contracts.new(...qv.getContractPayload({
+    const contract = zil.contracts.new(...qv.payloadQv({
         payload: {
             name: "Test hi",
             description: "Hello hi",
             options: ["opt1", "opt2", "opt3", "opt4"],
             creditToTokenRatio: "1000",
             //can register for next ~5 min
-            registrationEndTime: qv.getFutureTxBlockNumber(curBlockNumber, 60 * 5),
+            registrationEndTime: qv.futureTxBlockNumber(curBlockNumber, 60 * 5),
             //can vote in ~5 min and voting is open for ~10 min
-            expirationBlock: qv.getFutureTxBlockNumber(curBlockNumber, 60 * 15),
+            expirationBlock: qv.futureTxBlockNumber(curBlockNumber, 60 * 15),
             tokenId: "DogeCoinZilToken"
         }, ownerAddress: deployerAddress,
     }));
-    const [address, instance, deployTx] = await qv.deployContractHandle(
-        contract.deploy(...qv.getDeployPayload({ gasPrice }))
+    const [address, instance, deployTx] = await qv.handleDeploy(
+        contract.deploy(...qv.payloadDeploy({ gasPrice }))
     );
     console.log(address);
 
     /* Decentralized register (deployerAddress since it was set default in zil sdk)*/
-    const registerTx = await instance.call(...qv.registerPayload({
+    const registerTx = await instance.call(...qv.payloadRegister({
         gasPrice
     }));
     printReceipt(registerTx);
 
     /* Decentralized register (voterAddress)*/
     zil.wallet.setDefault(voterAddress);
-    const registerTx1 = await instance.call(...qv.registerPayload({
+    const registerTx1 = await instance.call(...qv.payloadRegister({
         gasPrice
     }));
     printReceipt(registerTx1);
